rename: 상단 스크롤 로직 유틸에서 훅으로 변경
